backend/requirements.txt
Flask==2.0.1
Flask-MySQL==1.5.2
Flask-Cors==3.0.10
python-dotenv==0.19.0
bcrypt==3.2.0
PyJWT==2.1.0
Werkzeug==2.0.3
flask-mysqldb==1.0.1